Add a Geojs geojson layer
This layer is similar to the annotation layer in that it is initialized
with a geojson object, but rendered features cannot be modified by the
layer itself.  The geojson prop provided to the layer is fully reactive
and will trigger a redraw for arbitrary changes to the geojson object
provided.

This component replaces all features on data changes; it is not designed
to be performant for large datasets.  Applications requiring frequent
redraw's and/or large numbers of features should modify the geojs
objects directly instead of relying on the built-in reactivity.