Add the ability to provide geojson styling
This adds a `featureStyle` prop to the geojson layer type.  Users can
provide an independent style for each of the raw geojson types (point,
line, and polygon).  The style object is passed directly to the
`feature.style` method, so anything supported by geojs can be used here.

Note: If a function is provided as a style property, then you may need
to manually call the `draw()` method on the layer component when it
needs to be rerendered.