feat(component): add a widget layer component
This adds a new GeojsWidgetLayer component for displaying arbitrary
content inside an element that pans with the map.  The logic in this
layer is primarily dedicated to setting the `left` and `top` css
properties of the widget parent when the map pans.

This default slot is appended to the layer's canvas on mount.  This
could cause problems when using with `v-if` directives depending on how
vue works.  I can make adjustments if there is ever a use case for this
behavior.

Props:
  * position: The lat/lon position to center the element on
  * size: The size of the component used to center the element correctly
  * offset: Shifts the anchor point by the given offset in pixels

Example use case: Adding a text label to the right of -100 longitude 40
latitude.

```pug
geojs-widget-layer(
  :position='{x: -100, y: 40}',
  :size='{width: 140, height: 50}',
  :offset='{x: 70, y: 0}'  // anchor to the left center
)
  p Label
```