feat(geojs): add zIndex and opacity props to all layers
This modifies the geojsLayer mixin to provide reactive zIndex and
opacity props to all layer types.  The opacity prop is now deduplicated
across all layers that previously supported the feature.

The zIndex prop is intend to provide a hook for developers to ensure the
layer order is consistent.  The previous behavior was that the z order
was determined by the order in which the layers were added to the map.
This causes issues for cases when the layer is added asynchronously into
the virtual DOM.