feat(girder): add a rest client for interacting with Girder
This commit adds a Girder "session" object which is a slightly extended
axios instance.  The session contains methods for handling
authentication and automatically inserts tokens into request headers.
This session object will be used by reuseable components and allow the
parent application to attach handlers to control the details of the
client/server communication.