refactor(geojs): use Vue's provide/inject API in the geojs layer mixin
This provides more flexibility in using the components compared to
accessing the `$parent` component directly.  This is particularly
relevant because newer versions of `@vue/test-utils` broke how we were
injecting the parent components in unit testing.

Note: This also updates `@vue/test-utils`, but not to the latest version
because there is a bug since `beta-18` causing props not to be updated.
This may (or may not) be related to the following issue:

https://github.com/vuejs/vue-test-utils/issues/738