feat(girder): add girder authentication dialogs
This adds a series of dialogs and forms that manage user authentication
as well as a new example to house girder related components.  These
components can be considered experimental versions of what could be
moved into the new girder web client.

The dialogs added are generated from lower-level forms that could be
used in page body (e.g. not in a dialog).  The forms have some client
side validation using Vee-Validate where possible.  When an error
error is returned from the server after submission, the `error` model
is updated with the message for the parent component to display.

There are some useability and design problems to be addressed in the
future as noted below.

Forms
=====

* GirderLoginForm:
  Log in an existing user.  Also contains links to the registration
  and reset password forms.  This should likely be extended to support
  OAuth something.

* GirderLogoutForm:
  Log out an existing user.  This could probably use some redesign.  I
  was imagining extending it to something like a "user settings dialog",
  but for now, it just a simple confirmation dialog.

* GirderRegisterForm:
  Register a new user.  Also contains a link to the login form.

* GirderResetPasswordForm:
  Reset a forgotten password.  This should be changed to display a
  message to check email rather than just closing on success.

Other
=====

* GirderUserDialog:
  This is a "meta" component that combines all of the forms and handles
  inter-linking between them for the most common use cases.

* GirderUserButton:
  This is a responsive button that changes depending on the user state.
  When not logged in, it is just a "LOGIN" button, otherwise it is a
  `v-avatar` component showing the user's initials.  We could do
  something more fancy like searching for gravatar images, etc.  This
  could use some redesign to make it more obvious how to register a new
  user.

As an additional note, the testing of the forms is a little ugly due to
bugs in synchronous behavior `@vue/test-utils`.  These are known
upstream and are targeted to be fixed after Vue@2.6 releases.

For this to work, I needed some of the changes in the more recent
versions of vuetify.  This also addresses issue #44.

fix #44