feat(geojs): make map viewport prop reactive
This change allows users of the GeojsMapViewport to request a viewport
(map center, zoom, rotation) by mutating the `viewport` prop.

This implementation is not perfect, but should satisify most use cases
without major performance implications.  Due to constraints on the
viewport (via clamping), existing user interactions, transitions, and
numerical artifacts, the requested viewport might not be honored.  In
this case, the `update:viewport` event will *not* be retriggered with
the real value.  This is done to avoid infinite update loops that would
have to be handled by the user of the component.

If there are issues with this approach, we could add a new event that
fires synchronously with the "real" viewport that is not attached to a
prop.

Fixes #43