feat(girder): add a job list view and associated components
This adds a (very preliminary) version of a job list component that
communicates with Girder's rest api.  All of the functionality of the
view is present, but there are currently no events propagated back up the
parent component.

There are a few things that could be done depending on the use cases and
priorities:
* Add more color and icons for statuses.
* Replace the job title with a clickable link that fires an event
* Expand and show more job details when clicking on a row
* Make an option for the rows to be selectable.
* Watch the EventStream (or poll) for job updates.

Fixes #52